Exercise_One/requirements.txt
Flask==1.1.2 
Jinja2==2.11.3
Werkzeug==1.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
SQLAlchemy==1.4.25
Flask-Bcrypt==1.0.1
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==4.0.0
Flask-SocketIO==5.1.1
pycryptodome==3.20.0
firebase-admin==5.0.1
python-dotenv==0.19.2
mysql-connector-python==8.0.5